[commhistory-daemon] Ignore duplicate MMS notifications. Contributes to JB#18573
If the user doesn't download the MMS message right away, operators
often resend the push notification. These were treated as separate
messages and littered the conversation view.

Now we use mmsId field to store the location url from the MMS push
notification. It's cleared as soon as the message gets downloaded.
If we receive a push message while another MMS with the same URL
is waiting for download, we ignore the duplicate.

Requires a few changes in mms-engine as well. Both old and new variants
of org.nemomobile.MmsHandler.messageNotification are supported (for now).
The new one includes the location URL, so that commhistoryd doesn't have
to parse the MMS PDU.