[commhistoryd] Don't use getEventById() result if no events are found. Contributes to JB#42099
getEventById() returns true if the database call succeeded, and
not necessarily whether a valid event was found.

The main issue here is that TextChannelListener::slotMessageSent()
was using an invalid event returned by getEventById(), filling that
with valid data and then failing on the saveMessage() call because
the event is actually invalid, and so the event with the updated
messageToken, status etc. is never saved to the db.

Most other calls to getEventById() already check whether it found
a valid event. The patch updates some other calls in MmsHandler
that were not checking this, to make sure those don't cause any
unintended side effects.