[commhistoryd] Refine group matching when there are multiple recipients. Contributes to JB#42099
When slotGroupInserted() is called, prefer to assign the current group
based on an exact recipient match rather than matching the first group
that contains any matching recipients.

This was already implemented for slotOnModelReady() by commit
1084b380e9d7870f736f0582bd4b4f28f223ae13 but needs to be done for
slotGroupInserted() as well.

This fixes an issue where if there are multiple message recipients, it
matches the group id for the broadcast message rather than the
individual message. Thus, the wrong group id is used when searching
the database for corresponding events, and so events cannot be located
and updated as expected.