[commhistoryd] Port to Python 3. JB#45864
Most changes come from 2to3 conversions, except for the comparator
function change and the buffer writing. Now uses sys.stdout.buffer.write()
instead of sys.stdout.write() so that the data is written directly to
the buffer instead of trying to consider any text encoding.

python.sh is no longer required as it is python2-specific.