[commhistory-daemon] Don't set x-nemo-max-content-lines notification hint. JB#50243
x-nemo-max-content-lines is currently set '1' to ensure the body text
is not shown for the notification. This hint has been deprecated, so
instead of setting that, just avoid setting a body and previewBody
instead.

SMS and MMS notifications should still show the body content, as those
previously had x-nemo-max-content-lines=6 to ensure the body was shown.